[lipstick] Simplify thermal notifier. JB#60190
This was quite much unnecessarily complicated, the thermal status getter
wasn't working properly, it was a blocking d-bus call, and the
whole getter was pointless when the code has been following all the
state changes anyway.

Also there's no point in following display state on the notifier
since it's creating critical notifications which turn on the screen.

The unit test was somewhat garbage too.
