+++ tests/ut_shutdownscreen/ut_shutdownscreen.pro
@@ -16,8 +16,6 @@ SOURCES += \
     $$STUBSDIR/homeapplication.cpp \
     $$DEVICESTATE/devicestate.cpp \
     $$DEVICESTATE/displaystate.cpp \
-    $$DEVICESTATE/thermal.cpp \
-    $$DEVICESTATE/ipcinterface.cpp \
     ut_shutdownscreen.cpp
 
 HEADERS += \
@@ -32,8 +30,5 @@ HEADERS += \
     $$DEVICESTATE/devicestate_p.h \
     $$DEVICESTATE/displaystate.h \
     $$DEVICESTATE/displaystate_p.h \
-    $$DEVICESTATE/thermal.h \
-    $$DEVICESTATE/thermal_p.h \
-    $$DEVICESTATE/ipcinterface_p.h \
     $$STUBSDIR/nemo-devicelock/devicelock.h \
     ut_shutdownscreen.h
