--- tests/ut_thermalnotifier/.gitignore
@@ -1,1 +1,0 @@
-ut_thermalnotifier
